Redirect by redirectPath method after logout
Change redirection of logout method to use redirectPath method. Redirection after logout can be set without need to overwrite entire method and same way as with login.
This could be especially helpful when multilangual routes are used so there are multiple routes to homepage and simple '/' as redirection path after logout is not sufficient.